third_party/WebKit/LayoutTests/media/video-preload-none-no-stalled-event-expected.txt

Test to see if a media element with preload=none dispatches a 'stalled' event.


EVENT(suspend)
END OF TEST
